% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearRatio.R
\name{bestGearRatio}
\alias{bestGearRatio}
\title{A Gear Function}
\usage{
bestGearRatio(m, maxDepth = 10, maxGear = 127, maxError = 0.1)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE}
}
\description{
This function allows you to express your love of cats.
}
\examples{
bestGearRatio()
}
\keyword{'continuous}
\keyword{'gear}
\keyword{fraction'}
\keyword{ratio'}
