% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearRatio.R
\name{bore3P}
\alias{bore3P}
\title{A Gear Function}
\usage{
bore3P(d1, d2, d3)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
bore3P()
}
\keyword{'continuous}
\keyword{'gear}
\keyword{fraction'}
\keyword{ratio'}
