% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CuttingSpeeds}
\alias{CuttingSpeeds}
\title{Cutting Speeds Table}
\format{A numeric data frame.}
\source{
Generated from \code{
Material <- c(  "Steel (tough)", 
               "Mild Steel", 
               "Mild Steel (with colant)", 
               "Cast Iron",
               "Alloy Steels (1320–9262)",
               "Carbon Steels (C1008–C1095)",
               "Free Cutting Steels (B1111–B1113 & C1108–C1213)",	
               "Stainless Steels (300 & 400 series)",	
               "Bronzes",	
               "Steel 12L14",	
               "Aluminium",	
               "Brass",	
               "Machinable Wax",	
               "Delrin",	
               "Polyethylene",	
               "Acrylic (with coolant)",	
               "Wood"
        )	
MPM.Min  <- c(18,  3,   6,  1, 3,  4,  35,  23, 24, 9,  122,  90,  6,  11, 12, 15, 183) 
MPM.Max  <- c(50,  38, 	7,	2, 20, 21, 69,	40,	45,	10, 305,  210, 6,  11, 12, 15, 305)	
SFM.Min  <- c(60,  10,  20, 6, 12, 10, 115, 30, 10,	30, 400,  300, 20, 35, 40, 50, 600)
SFM.Max  <- c(100, 125, 25, 8, 65, 70, 225, 7,  80, 33, 1000, 700, 20, 35, 40, 50, 1000)

cuttingSpeeds <- data.frame(Material, MPM.Min, MPM.Max, SFM.Min, SFM.Max, stringsAsFactors = FALSE)
}
}
\description{
Cutting speeds for various materials  
using a plain high speed steel cutter
}
\examples{
data(cuttingSpeeds)
str(cuttingSpeeds)
}
