% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearRatio.R
\name{gearRatio}
\alias{gearRatio}
\title{A Gear Function}
\usage{
gearRatio(m, n = 1, depth = 20)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
gearRatio()
}
\keyword{'continuous}
\keyword{'gear}
\keyword{fraction'}
\keyword{ratio'}
