% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MetricTapDrills}
\alias{MetricTapDrills}
\title{Starrett Metric Tap Drill Sizes}
\format{A numeric data frame.}
\source{
Generated from \code{
ThreadSizeMetric <- c(
                       1.6,   1.8,   2  ,   2.2,   2.5,   3,
                       3.5,   4  ,   4.5,   5  ,   6  ,   7,
                       8  ,   8  ,  10  ,  10  ,  12  ,  12,
                      14  ,  14  ,  16  ,  16  ,  18  ,  18,
                      20  ,  20  ,  22  ,  22  ,  24  ,  24,
                      27  ,  27  ,  30  ,  30  ,  33  ,  33,
                      36  ,  36  ,  39  ,  39
                      )
ThreadPitchMeric <- (
                     0.35,  0.35,  0.4 ,   0.45,  0.45,  0.5 ,
                     0.6 ,  0.7 ,  0.75,   0.8 ,  1   ,  1   ,
                     1.25,  1   ,  1.5 ,   1.25,  1.75,  1.25,
                     2   ,  1.5 ,  2   ,   1.5 ,  2.5 ,  1.5 ,
                     2.5 ,  1.5 ,  2.5 ,   1.5 ,  3   ,  2   ,
                     3   ,  2   ,  3.5 ,   2   ,  3.5,   2   ,  
                     4   ,  3   ,  4   ,   3
                     )
TapDrillMetric <- (
                    1.25,   1.45,   1.60,   1.75,   2.05,   2.50,
                    2.90,   3.30,   3.70,   4.20,   5.00,   6.00,
                    6.70,   7.00,   8.50,   8.70,  10.20,  10.80,
                   12.00,  12.50,  14.00,  14.50,  15.50,  16.50,
                   17.50,  18.50,  19.50,  20.50,  21.00,  22.00,
                   24.00,  25.00,  26.50,  28.00,  29.50,  31.00,
                   32.00,  33.00,  35.00,  36.00
                   )
TapDrillInch <- c(
                 0.0492, 0.0571, 0.0630, 0.0689, 0.0807, 0.0984, 
                 0.1142, 0.1299, 0.1457, 0.1654, 0.1968, 0.2362,
                 0.2638, 0.2756, 0.3346, 0.3425, 0.4016, 0.4252,
                 0.4724, 0.4921, 0.5512, 0.5709, 0.6102, 0.6496,
                 0.6890, 0.7283, 0.7677, 0.8071, 0.8268, 0.8661,
                 0.9449, 0.9843, 1.0433, 1.1024, 1.1614, 1.2205,
                 1.2598, 1.2992, 1.3780, 1.4173
                 )
 MetricTapDrills <- data.frame(ThreadSizeMetric,
                               ThreadPitchMetric,
                               TapDrillMetric, 
                               TapDrillInch, 
                               stringsAsFactors = FALSE)
}
}
\description{
A few numbers from \code{rnorm()}.
}
\examples{
data(InchTapDrills)
str(InchTapDrills)
}
