% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearRatio.R
\name{plotCircle}
\alias{plotCircle}
\title{A Gear Function}
\usage{
plotCircle(
  radius = 1,
  main = "",
  centerX = 0,
  centerY = 0,
  color = "black",
  npoints = 100
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
plotCircle()
}
\keyword{'continuous}
\keyword{'gear}
\keyword{fraction'}
\keyword{ratio'}
