% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearRatio.R
\name{print.gearRatioList}
\alias{print.gearRatioList}
\title{A Gear Function}
\usage{
\method{print}{gearRatioList}(x, ...)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
print.gearRatioList()
}
\keyword{'continuous}
\keyword{'gear}
\keyword{fraction'}
\keyword{ratio'}
