% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedSpeed.R
\name{spindleSpeed}
\alias{spindleSpeed}
\title{A Gear Function}
\usage{
spindleSpeed(SFM, toolDiameter, units = "Metric")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
spindleSpeed()
}
\keyword{'continuous}
\keyword{'gear}
\keyword{fraction'}
\keyword{ratio'}
