% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StarrettTapDrills}
\alias{StarrettTapDrills}
\title{Starrett Tab Drill Sizes}
\format{A numeric data frame.}
\source{
Generated from \code{set.seed(123); rnorm(10)}
}
\description{
A few numbers from \code{rnorm()}.
}
\examples{
data(StarrettTapDrills)
str(StarrettTapDrills)
}
